requirements.txt
Django==1.8.2
-e git+git@github.com:talentsprout/PorpoiseFlow.git@master#egg=PorpoiseFlow-master
coverage==3.7.1
django-extensions==1.5.3
django-filter==0.9.2
djangorestframework==3.1.1
djangorestframework-xml==1.0.1
djangorestframework-yaml==1.0.2
gnureadline==6.3.3
inflection==0.3.1
ipdb==0.8.1
ipython==3.2.1
pydot==1.0.2
pygraphviz==1.2
pyparsing==1.5.7
requests==2.7.0
rest-framework-ember==1.3.2
six==1.9.0
wsgiref==0.1.2
